Added Inventory validation in PO service + refactor Handlers in Article Service for cleaner code

+++ MiniERP.PurchaseOrderService/Dtos/PODetailCreateDto.cs
@@ -4,8 +4,8 @@
     {
         public int LineNo { get; set; }
         public int Productd { get; set; }
-        public decimal Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
+        public double Quantity { get; set; }
+        public double UnitPrice { get; set; }
         public string Notes { get; set; } = string.Empty;
     }
 }
